Develop Fzf-pick program to pick records
This program shall read records from the standard input stream, present
them in an interactive search screen to the user and wait for her/him to
fuzzy-find and pick the records s/he is interested in.  Once the user
made her/his selection _Fzf-pick_ shall write the selected records to
the standard output stream.

When given the `help` option _Fzf-pick_ shall print its help text and
exit.

This program shall print one record per line and quote it when connected
to a terminal for readability.  But when its output is redirected, e.g.
piped through another program for further processing, it shall print
unquoted but null-terminated records instead to make them securely
processible.

Signed-off-by: Tim Friske <me@tifr.de>