Call Fzf-pick's function to pick records
This function reads records from the standard input stream, presents
them in an interactive search screen to the user and waits for her/him
to fuzzy-find and pick the records s/he is interested in.  Once the user
made her/his selection _Fzf-pick_ writes the selected records to the
standard output stream.

Signed-off-by: Tim Friske <me@tifr.de>