Let Fzf-pick print command line w/ records picked
_Fzf-pick_ should be usable in scenarios where it cannot execute the
given command itself because it runs as a child process and that command
can only be successfully executed by its parent process.

For example _Bash's_ built-in command `cd` for changing directories can
only be invoked by the user's shell directly in order to have it change
the user's working directory.

One way to circumvent this limitation is to let _Fzf-pick_ print the
given command with the records selected as arguments.  The resulting
command line can then be evaluated by the parent process, e.g. the
user's shell, in order to have an effect.

Signed-off-by: Tim Friske <me@tifr.de>