Let Fzf-pick print line w/ cmd and records as args
Make the _Fzf-pick_ program usable in scenarios where the command given
to it should have an effect on its parent process and therefore that
command cannot be executed by the child process itself, namely
_Fzf-pick_.  For example the current working directory of the user's
shell can only be changed if the shell built-in command `cd` is executed
directly by that shell.  The same rules apply when it comes to _Bash's_
job control commands `jobs`, `fg` and `bg`.

Signed-off-by: Tim Friske <me@tifr.de>