Let Fzf-pick print either quoted or null-term recs
When _Fzf-pick_ is not connected to a terminal and `quoting` is not
requested let it print raw but null-terminated records instead to the
standard output stream.  The latter is the more robust format if records
contain special characters because the only special character that must
*not* appear in the records is the `\0` null character.  The `quoted'
format on the other hand is only understood by a few programs, e.g.
shells such as _Bash_.

Signed-off-by: Tim Friske <me@tifr.de>