Develop _Fzf-quote_ program to quote special chars
The program _Fzf-quote_ shall read unquoted records from the standard
input stream, quote all special characters it finds in these records and
write them in quoted form back to the standard output stream.

When _Fzf-quote's_ standard input stream is connected to a terminal it
shall expect each unquoted record to appear on a new line.  Otherwise,
e.g. when _Fzf-quote_ reads from a file or another program, it shall
assume each unquoted record to be terminated with a `\0` null character.
The latter assumption shall be overridable by passing _Fzf-quote's_
`read` variable set to either `1` or `true`.

When _Fzf-quote's_ standard output stream is connected to a terminal it
shall write each quoted record on a new line.  Otherwise, for example
when _Fzf-quote_ writes to a file or another program, it shall terminate
each quoted record with a `\0` null character.  The latter decision
shall be overridable by passing _Fzf-quote's_ `write` variable set to
either `1` or `true`.

Signed-off-by: Tim Friske <me@tifr.de>