Let Fzf-quote write line- and null-term records
When writing data to a terminal its records usually come in lines since
terminals are line-oriented per se.  But when writing data to a file or
another program chances are its records come in the form of `\0` null-
terminated records since this format allows a record to contain all
characters except the `\0` null character.

Signed-off-by: Tim Friske <me@tifr.de>