Let Fzf-pick correctly print given command w/ args
When invoking another _Bash_ instance with a command and arguments to
execute then this command and its arguments must be passed as one
string.

Before — without quoting — only the command but not its arguments were
executed.  This lead to the following error message; for example when
passing `echo` as the command to prepend to the picked records:

----
$ cmd='echo' fzf-pick
printf: usage: printf [-v var] format [arguments]
----

Signed-off-by: Tim Friske <me@tifr.de>