Find files with hotkey `Ctrl-ad` and Fzf-folders
When the user presses the `Ctrl-ad` hotkey list the folders below the
current directory in _Fzf's_ interactive search screen for the user to
fuzzy-find and pick the folders s/he is interested in.  Once selected
_Fzf-insert_ inserts these folders into the command line editing buffer of
the user's shell at the cursor's current position.

Try to source the _Fzf-folders_ _Bash_ script.  Do not complain if this
script is not present.  Treat it as an optional feature and therefore do
not depend on it.

Signed-off-by: Tim Friske <me@tifr.de>