Give main function same name as file type utility
As a matter of programming style and for consistency with the other
programs written by us rename the main function `file-type` as
`file_type`.

Note the subtle difference of using underscores in the function name as
opposed to hyphens in the program name.  Hypens should be easier to type
by the user when frequently invoking this program.  But stick with
underscores in identifiers since most programming languages do not
support hyphens in them.

Signed-off-by: Tim Friske <me@tifr.de>