Collect key bindings to call Fzf program with
Use an array to incrementally build _Fzf's_ key bindings.  An array can
also be better converted into an arbitrarily separated string of values,
e.g.  with `"example_array[@]"` into a space-separated string of values.

Hard-code all of the key bindings under the assumption that they are
`good' enough and do not need to be configurable by external means.

Signed-off-by: Tim Friske <me@tifr.de>