Resolve symlinks in paths and deduplicate them
Keep the list of folders short by appending only folders which are not
yet contained in that list.  By resolving symbolic links chances are
that two otherwise distinct folders are the same.  This might reduce the
number of folders even further.

On modern _Linux_ distributions some of the folders that used to contain
specific binaries and scripts are nowadays merely pointers to a few
larger-sized folders which remained.  But for compatibility reasons the
old locations are kept but got replaced with symlinks to the few
remaining folders.

Signed-off-by: Tim Friske <me@tifr.de>