Differentiate between user and system setup
Configuring a user as well as the system s/he is working on at the same
time does not work well.  All user specific configuration files and
directories should be set up with the user as the owner and her/his
primary group as the group.  On the other hand setting up the system
wide configuration files and directories requires more privileges such
as performing operations as another user, usually the `root' user.

Also the frequency with which user- and system specific configuration
files are changed and thus require re-configuration substantially
differ; not only by the number of files but also the number of users
altering their files on a regular basis.

All of the above points led to the conclusion that user and system
setup are two distinct things and therefore should be performed
separately.

Signed-off-by: Tim Friske <me@tifr.de>