Redefine Mktar Bash script as alias
Due to the multitude of options how to create a _TAR_ _tape archive_
with the _Tar_ program it turned out that it is simply not feasible to
cover all of the practical use cases.  After some iterations of trying
to implement all of these in a compatible way it is decided to replace
the _Mktar_ _Bash_ script and companions with a couple of simple _Bash_
aliases.

Signed-off-by: Tim Friske <me@tifr.de>