Allow passing args to disk usage prog in aliases
In the old version of the _Bash_ aliases for reporting disk usage of
files (duf) and directories (dud) they always reported that information
about the current working directory.

In the new version the user can now pass any arguments to the disk usage
program, such as the files and directories s/he wants to be informed
about.

Signed-off-by: Tim Friske <me@tifr.de>