Make sure to list all files with Bash aliases
For _Bash's_ glob pattern `**` to be recognized its `globstar` option
must be enabled.  Do not depend on this option to be already enabled
because it is disabled by default.

Therefore enable this option for the duration of listing the files with
_Bash_ aliases in order to discover the files not only in the given
directory but also in all of all of its subdirectories.

Signed-off-by: Tim Friske <me@tifr.de>