Replace Git Add aliases w/ commands
Instead of having them defined as aliases inside of one of _Git's_
configuration files they should each be implemented as a command in its
own executable file for _Git_ to look it up only when needed.  In other
words the previous aliases are regular programs now which can be found
in one of the folders listed in the `PATH` environment variable, e.g.
below the `~/.local/bin` folder.

Signed-off-by: Tim Friske <me@tifr.de>