Add shell script to set Tmux working directory
This shell script is executed each time a new prompt is displayed.  For
that to happen the `PROMPT_COMMAND` environment variable is adapted.

It helps Tmux to open panes in the current working directory even if
that folder is a symbolic link when new windows are created or existing
are splitted.

Signed-off-by: Tim Friske <me@tifr.de>