Add fzf-fd utility program
The `fzf-fd` utility program lists the files and folders below the
current directory with the help of the `fd` program and pipes the
listing through the `fzf` filter program. The user can interactively
filter and select the files or folders s/he is eventually interested in.

This utility program works similary to the already existing `fzf-find`.
Both programs share the same command line interface, e.g. by accepting
the same set of parameters.

The `fzf-finder` utility program is simply a soft link to either one of
`fzf-fd` or `fzf-find` depending on whether the `fd` executable is in
the `PATH`.

Signed-off-by: Tim Friske <me@tifr.de>