Strip heads/ prefix from branch listings
Since the `git-for-each-ref` program deals with references in general
which could be branches, tags etc. it prints the prefixes `heads/`,
`tags/` etc. respectively in order to disambiguate two references which
would otherwise share the same name.

In the context of the `git-bg` program only branches are printed. Thus
remove the `heads/` prefix which is redundant in this case.

Signed-off-by: Tim Friske <me@tifr.de>