Don't abort if Git config option value not set
The `git-you` program retrieves a previously set option value and uses
that as a default value when prompting the user for a new value.

After following the ShellCheck advice in another commit to separate the
declaration of local variables from their assignments in order to not
mask exit codes of command substitutions the introduced changes were not
tested. This led `git-you` to abort execution in case there was no
previously set config option value.

Also, optionally print the optional config subsection in parenthesis in
the user input prompts instead of just prepending it to the section.

Signed-off-by: Tim Friske <me@tifr.de>