Make k8s-suspend-cronjobs more robust
Move the command `kubectl patch` and the variable `suspension_toggle`
into the exported Bash function `toggle_suspension` which gets invoked
by `parallel` instead. This helps to avoid the backslash-escapes in the
variable that are needed otherwise because the value would get twice
interpreted; once by passing it to the `kubectl patch` command and a
second time when `parallel` invokes that command.

Signed-off-by: Tim Friske <me@tifr.de>